Trying to prevent hacker uploading a malicious file
As Eldy alerted, there is a risk of someone first uploading a FAKE and 
malicious `main.inc.php` file and then managing to run it, using the 
web browser form that appears in case it is the first time we use the 
third-party module and the script is not be able to find the 
`main.inc.php` file.

It's true that the risk is very minimal because:

1. The script `main_module.inc.php` only displays on the web browser the form to write the file path of `main.inc.php` when this file cannot be found in the usual locations. So this affects only a very low percentage of users.
2. This webform appears ONLY ONCE for a Dolibarr instance: when installing a third-party module. So again, it's a very low percentage of cases where a hacker is able to know the exact day and hour of this installation by the admin user of the Dolibarr instance. 

Anyway, I've added these 2 extra checks:

1. The filename must be `main.inc.php` (this is obvious, but it wasn't being checked).
   
2. The parent directory containing this file must also contain one of 
the core directories of the Dolibarr core and its sub-directories 
(`fourn/facture/tpl`). So for the hack to be successful, the hacker needs to be able also to create directories and subdirectories on the server... which is also very unlikely if that user does not have access to FTP/SSH.

Although it could still be possible. So please, suggest new defenses :-)